requirements.txt
asgiref==3.4.1
astroid==2.9.3
autopep8==1.6.0
Brotli==1.0.9
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
cssselect2==0.5.0
dj-database-url==0.5.0
Django==3.2.5
django-braces==1.14.0
django-ckeditor==6.1.0
django-crispy==1.8.1
django-crispy-forms==1.12.0
django-crontab==0.7.1
django-embed-video==1.4.0
django-js-asset==1.2.2
django-memcache-status==2.2
django-rosetta==0.9.6
django-taggit==1.5.1
djangorestframework==3.12.4
easy-thumbnails==2.7.1
flake8==4.0.1
fonttools==4.31.2
html5lib==1.1
idna==3.2
isort==5.10.1
jedi==0.18.1
lazy-object-proxy==1.7.1
Markdown==3.3.4
mccabe==0.6.1
parso==0.8.3
Pillow==8.3.1
platformdirs==2.5.1
pluggy==1.0.0
polib==1.1.1
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pydyf==0.1.2
pyflakes==2.4.0
pylint==2.12.2
pyphen==0.12.0
python-dotenv==0.19.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.4.0
python-memcached==1.59
pytz==2021.1
reportlab==3.6.8
requests==2.26.0
rope==0.23.0
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.1
tinycss2==1.1.1
toml==0.10.2
ujson==5.1.0
unicode==2.8
unicode-slugify==0.1.3
Unidecode==1.2.0
urllib3==1.26.6
weasyprint==54.2
webencodings==0.5.1
wrapt==1.13.3
yapf==0.32.0
zopfli==0.2.1
